# perf: refactor messages to reduce a duplicate serialization