--- a/ironfish/src/network/messages/cannotSatisfyRequest.ts
+++ b/ironfish/src/network/messages/cannotSatisfyRequest.ts
@@ -9,8 +9,8 @@ export class CannotSatisfyRequest extends RpcNetworkMessage {
     super(NetworkMessageType.CannotSatisfyRequest, Direction.Response, rpcId)
   }
 
-  serialize(): Buffer {
-    return Buffer.alloc(0)
+  serializePayload(): void {
+    return
   }
 
   static deserialize(rpcId: number): CannotSatisfyRequest {