--- a/ironfish/src/network/messages/getBlockTransactions.test.ts
+++ b/ironfish/src/network/messages/getBlockTransactions.test.ts
@@ -1,7 +1,12 @@
 /* This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
-import { createNodeTest, useMinersTxFixture, useTxSpendsFixture } from '../../testUtilities'
+import {
+  createNodeTest,
+  serializePayload,
+  useMinersTxFixture,
+  useTxSpendsFixture,
+} from '../../testUtilities'
 import { expectGetBlockTransactionsResponseToMatch } from '../testUtilities'
 import {
   GetBlockTransactionsRequest,
@@ -15,7 +20,7 @@ describe('GetBlockTransactionsRequest', () => {
     const transactionIndexes = [1, 60000]
 
     const message = new GetBlockTransactionsRequest(blockHash, transactionIndexes, rpcId)
-    const buffer = message.serialize()
+    const buffer = serializePayload(message)
     const deserializedMessage = GetBlockTransactionsRequest.deserialize(buffer, rpcId)
 
     expect(deserializedMessage).toEqual(message)
@@ -34,7 +39,7 @@ describe('GetBlockTransactionsResponse', () => {
     const transactions = [transactionA, transactionB]
 
     const message = new GetBlockTransactionsResponse(blockHash, transactions, rpcId)
-    const buffer = message.serialize()
+    const buffer = serializePayload(message)
     const deserializedMessage = GetBlockTransactionsResponse.deserialize(buffer, rpcId)
 
     expectGetBlockTransactionsResponseToMatch(message, deserializedMessage)