--- a/ironfish/src/network/messages/identify.ts
+++ b/ironfish/src/network/messages/identify.ts
@@ -62,8 +62,7 @@ export class IdentifyMessage extends NetworkMessage {
     this.features = features
   }
 
-  serialize(): Buffer {
-    const bw = bufio.write(this.getSize())
+  serializePayload(bw: bufio.StaticWriter | bufio.BufferWriter): void {
     bw.writeBytes(Buffer.from(this.identity, 'base64'))
     bw.writeVarString(this.name, 'utf8')
     bw.writeU16(this.port)
@@ -78,8 +77,6 @@ export class IdentifyMessage extends NetworkMessage {
     let flags = 0
     flags |= Number(this.features.syncing) << 0
     bw.writeU32(flags)
-
-    return bw.render()
   }
 
   static deserialize(buffer: Buffer): IdentifyMessage {