--- a/ironfish/src/network/messages/newBlockHashes.test.ts
+++ b/ironfish/src/network/messages/newBlockHashes.test.ts
@@ -1,6 +1,7 @@
 /* This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
+import { serializePayload } from '../../testUtilities'
 import { NewBlockHashesMessage } from './newBlockHashes'
 
 describe('NewBlockHashesMessage', () => {
@@ -16,7 +17,7 @@ describe('NewBlockHashesMessage', () => {
       },
     ])
 
-    const buffer = message.serialize()
+    const buffer = serializePayload(message)
     const deserializedMessage = NewBlockHashesMessage.deserialize(buffer)
     expect(deserializedMessage).toEqual(message)
   })