--- a/ironfish/src/network/messages/newCompactBlock.ts
+++ b/ironfish/src/network/messages/newCompactBlock.ts
@@ -15,12 +15,8 @@ export class NewCompactBlockMessage extends NetworkMessage {
     this.compactBlock = compactBlock
   }
 
-  serialize(): Buffer {
-    const bw = bufio.write(this.getSize())
-
+  serializePayload(bw: bufio.StaticWriter | bufio.BufferWriter): void {
     writeCompactBlock(bw, this.compactBlock)
-
-    return bw.render()
   }
 
   static deserialize(buffer: Buffer): NewCompactBlockMessage {