--- a/ironfish/src/network/messages/signal.ts
+++ b/ironfish/src/network/messages/signal.ts
@@ -40,13 +40,11 @@ export class SignalMessage extends NetworkMessage {
     this.signal = signal
   }
 
-  serialize(): Buffer {
-    const bw = bufio.write(this.getSize())
+  serializePayload(bw: bufio.StaticWriter | bufio.BufferWriter): void {
     bw.writeBytes(Buffer.from(this.destinationIdentity, 'base64'))
     bw.writeBytes(Buffer.from(this.sourceIdentity, 'base64'))
     bw.writeBytes(Buffer.from(this.nonce, 'base64'))
     bw.writeBytes(Buffer.from(this.signal, 'base64'))
-    return bw.render()
   }
 
   static deserialize(buffer: Buffer): SignalMessage {