--- a/ironfish/src/workerPool/tasks/verifyTransactions.ts
+++ b/ironfish/src/workerPool/tasks/verifyTransactions.ts
@@ -14,13 +14,11 @@ export class VerifyTransactionsRequest extends WorkerMessage {
     this.transactionsPosted = transactionsPosted
   }
 
-  serialize(): Buffer {
-    const bw = bufio.write(this.getSize())
+  serializePayload(bw: bufio.StaticWriter | bufio.BufferWriter): void {
     bw.writeU64(this.transactionsPosted.length)
     for (const tx of this.transactionsPosted) {
       bw.writeVarBytes(tx)
     }
-    return bw.render()
   }
 
   static deserialize(jobId: number, buffer: Buffer): VerifyTransactionsRequest {
@@ -52,10 +50,8 @@ export class VerifyTransactionsResponse extends WorkerMessage {
     this.verified = verified
   }
 
-  serialize(): Buffer {
-    const bw = bufio.write(this.getSize())
+  serializePayload(bw: bufio.StaticWriter | bufio.BufferWriter): void {
     bw.writeU8(Number(this.verified))
-    return bw.render()
   }
 
   static deserialize(jobId: number, buffer: Buffer): VerifyTransactionsResponse {