## Summary

The 3 message types that implement the `Serializable` interface have essentially a 2-step serialization process. Step 1: serialize the metadata/header, and step 2: serialize the payload data.

We were instantiating a new bufio writer and calling `.render()` in both steps, despite step 1 calling step 2 and immediately writing the combined serialized message into a buffer.

By changing the payload serialization to take in an already instantiated bufio writer, we can write directly into the same writer, and not have to `.render()` it twice.

This should lead to substantial performance gains for both the worker pool and the peer network.

```
Performance change summary:

Stats taken from 50 iterations each
Staging -> This PR

createMinersFeeRequest:
Elapsed: 3.4945 ms -> 2.4316 ms (~30% faster)
Fastest: 0.0268 ms -> 0.0162 ms (~40% faster)
Average: 0.0699 ms -> 0.0486 ms (~30% faster)
Slowest: 0.921479 ms -> 0.6202705 (~33% faster)

decryptNotesRequest (100 notes):
Elapsed: 28.5183 ms -> 25.8990 ms (~9% faster)
Fastest: 0.309625 -> 0.3159585 (~2% slower)
Average: 0.57036673 -> 0.5179804 (~9% faster)
Slowest: 2.5901255 -> 1.563854 (~40% faster)
```

## Testing Plan

## Documentation

Does this change require any updates to the Iron Fish Docs (ex. [the RPC API
Reference](https://ironfish.network/docs/onboarding/rpc/chain))? If yes, link a
related documentation pull request for the website.

```
[ ] Yes
```

## Breaking Change

Is this a breaking change? If yes, add notes below on why this is breaking and
what additional work is required, if any.

```
[ ] Yes
```